requirements.txt
Django==1.3
celery==2.2.5
django-celery==2.2.4
-e git://github.com/toastdriven/django-haystack.git@9d75a9c031c845201ddf40e75824b4613b8c369e#egg=django_haystack-1.2.0_beta-py2.7-dev
django-mailer==0.1.0
pysolr==2.0.13
pytz==2011e
lxml==2.3

